+++ src/index.ts
@@ -2,6 +2,6 @@ export * from "./safeHandler.js";
 export * from "./camelize.js";
 export * from "./errors.js";
 export * as validator from "./validator.js";
-export * as fs_utils from "./fs.js";
+export * as fs from "./fs.js";
 export * from "./types.js";
 export { default as convert } from "./convert.js";